Validering ved opprettelse av medlemskapsperioder. Endepunkt for å returnere folketrygd-bestemmelse med tilhørende vilkår

+++ service/src/main/java/no/nav/melosys/service/medlemskapsperiode/MedlemskapsperiodeService.java
@@ -2,18 +2,16 @@ package no.nav.melosys.service.medlemskapsperiode;
 
 import java.time.LocalDate;
 import java.util.Collection;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import no.nav.melosys.domain.Behandling;
-import no.nav.melosys.domain.Behandlingsresultat;
-import no.nav.melosys.domain.InnvilgelsesResultat;
-import no.nav.melosys.domain.Medlemskapsperiode;
+import no.nav.melosys.domain.*;
 import no.nav.melosys.domain.behandlingsgrunnlag.Behandlingsgrunnlag;
 import no.nav.melosys.domain.behandlingsgrunnlag.SoeknadFtrl;
-import no.nav.melosys.domain.kodeverk.Folketrygdloven_kap2_bestemmelser;
-import no.nav.melosys.domain.kodeverk.Medlemskapstyper;
-import no.nav.melosys.domain.kodeverk.Trygdedekninger;
+import no.nav.melosys.domain.kodeverk.*;
 import no.nav.melosys.exception.FunksjonellException;
 import no.nav.melosys.exception.IkkeFunnetException;
 import no.nav.melosys.exception.MelosysException;
@@ -23,13 +21,16 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import static com.google.common.collect.MoreCollectors.onlyElement;
+import static java.lang.String.format;
+import static no.nav.melosys.domain.kodeverk.Trygdedekninger.FULL_DEKNING_EOSFO;
+import static no.nav.melosys.domain.kodeverk.Trygdedekninger.FULL_DEKNING_FTRL;
 import static no.nav.melosys.service.kontroll.PeriodeKontroller.feilIPeriode;
 
 @Service
 public class MedlemskapsperiodeService {
 
     private static final Collection<Trygdedekninger> GYLDIGE_TRYGDEDEKNINGER = Stream.of(Trygdedekninger.values())
-        .filter(trygdedekning -> trygdedekning != Trygdedekninger.FULL_DEKNING_EOSFO).collect(Collectors.toSet());
+        .filter(trygdedekning -> trygdedekning != FULL_DEKNING_EOSFO && trygdedekning != FULL_DEKNING_FTRL).collect(Collectors.toSet());
 
     private final MedlemskapsperiodeRepository medlemskapsperiodeRepository;
     private final BehandlingsresultatService behandlingsresultatService;
@@ -118,8 +119,14 @@ public class MedlemskapsperiodeService {
     }
 
     @Transactional(rollbackFor = MelosysException.class)
-    public Collection<Medlemskapsperiode> utledMedlemskapsperioderFraSøknad(long behandlingID, Folketrygdloven_kap2_bestemmelser bestemmelse) throws IkkeFunnetException {
+    public Collection<Medlemskapsperiode> utledMedlemskapsperioderFraSøknad(long behandlingID, Folketrygdloven_kap2_bestemmelser bestemmelse) throws FunksjonellException {
+        if (!støtterBestemmelse(bestemmelse)) {
+            throw new FunksjonellException("Støtter ikke perioder med bestemmelse " + bestemmelse);
+        }
+
         Behandlingsresultat behandlingsresultat = behandlingsresultatService.hentBehandlingsresultat(behandlingID);
+        validerSakstype(behandlingsresultat.getBehandling().getFagsak());
+        validerVilkår(behandlingsresultat, bestemmelse);
         medlemskapsperiodeRepository.deleteByBehandlingsresultat(behandlingsresultat);
 
         Behandling behandling = behandlingsresultat.getBehandling();
@@ -138,5 +145,40 @@ public class MedlemskapsperiodeService {
 
         medlemskapsperioder.forEach(m -> m.setBehandlingsresultat(behandlingsresultat));
         return medlemskapsperiodeRepository.saveAll(medlemskapsperioder);
+    }
+
+    private void validerVilkår(Behandlingsresultat behandlingsresultat, Folketrygdloven_kap2_bestemmelser bestemmelse) throws FunksjonellException {
+        var vilkårForBestemmelse = hentVilkårForBestemmelse(bestemmelse);
+        if (!behandlingsresultat.oppfyllerVilkår(vilkårForBestemmelse)) {
+            throw new FunksjonellException(format("Vilkår %s er påkrevd for bestemmelse %s", vilkårForBestemmelse, bestemmelse));
+        }
+    }
+
+    private void validerSakstype(Fagsak fagsak) throws FunksjonellException {
+        if (fagsak.getType() != Sakstyper.FTRL) {
+            throw new FunksjonellException("Kan ikke opprette medlemskapsperioder for sakstype " + fagsak.getType());
+        }
+    }
+
+    public Collection<Trygdedekninger> hentGyldigeTrygdedekninger() {
+        return GYLDIGE_TRYGDEDEKNINGER;
+    }
+
+    public Map<Folketrygdloven_kap2_bestemmelser, Collection<Vilkaar>> hentBestemmelserMedVilkaar() {
+        return Map.of(
+            Folketrygdloven_kap2_bestemmelser.FTRL_KAP2_2_8_FØRSTE_LEDD_A,
+            Set.of(Vilkaar.FTRL_2_8_FORUTGÅENDE_TRYGDETID),
+            Folketrygdloven_kap2_bestemmelser.FTRL_KAP2_2_8_ANDRE_LEDD,
+            Set.of(Vilkaar.FTRL_2_8_FORUTGÅENDE_TRYGDETID, Vilkaar.FTRL_2_8_NÆR_TILKNYTNING_NORGE)
+        );
+    }
+
+    private Collection<Vilkaar> hentVilkårForBestemmelse(Folketrygdloven_kap2_bestemmelser bestemmelse) throws FunksjonellException {
+        return Optional.ofNullable(hentBestemmelserMedVilkaar().get(bestemmelse))
+            .orElseThrow(() -> new FunksjonellException("Finner ikke vilkår for bestemmelse " + bestemmelse));
+    }
+
+    private boolean støtterBestemmelse(Folketrygdloven_kap2_bestemmelser bestemmelse) {
+        return hentBestemmelserMedVilkaar().containsKey(bestemmelse);
     }
 }
